address feedback

+++ pkg/ipmi/aes_128_cbc.go
@@ -65,11 +65,11 @@ func (a *AES128CBC) DecodeFromBytes(data []byte, _ gopacket.DecodeFeedback) erro
 	// ciphertext is useless on its own, so we don't mind
 	mode.CryptBlocks(data[a.cipher.BlockSize():], data[a.cipher.BlockSize():])
 
-	padBytes := data[len(data)-1]
+	padBytes := uint8(data[len(data)-1])
 	// table 13-20 of the spec says the confidentiality pad length ranges from
 	// 0 to 15 bytes if using AES, but we may receive 16 bytes if the BMC's
-	// implementation of AES in CBC mode requires a minimum of one pad byte and
-	// the message is already aligned.
+	// implementation of AES in CBC mode requires a minimum of one pad byte
+	// (which is how OpenSSL works) and the message is already aligned.
 	if padBytes > uint8(a.cipher.BlockSize()) {
 		return fmt.Errorf("invalid number of pad bytes: %v", padBytes)
 	}
